# src/Services/FirmwareService/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/firmware/v1/campaigns": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Создает новую кампанию OTA-обновления для указанной прошивки со статусом Pending.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rollout Campaigns"
                ],
                "summary": "Создать кампанию",
                "parameters": [
                    {
                        "description": "Данные для создания кампании",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presentation.CreateCampaignRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Возвращает ID созданной кампании",
                        "schema": {
                            "$ref": "#/definitions/internal_presentation.IDResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/firmware/v1/campaigns/{id}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Полностью удаляет кампанию и все связанные с ней таргеты из базы данных (Каскадное удаление).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rollout Campaigns"
                ],
                "summary": "Удалить кампанию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID кампании",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Кампания успешно удалена"
                    },
                    "400": {
                        "description": "Неверный ID кампании",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/firmware/v1/campaigns/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Переводит кампанию в статус Canceled (Аварийная остановка развертывания из-за багов или ошибок).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rollout Campaigns"
                ],
                "summary": "Отменить кампанию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID кампании",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Указание причины отмены",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presentation.CancelCampaignRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Кампания успешно отменена"
                    },
                    "400": {
                        "description": "Неверный формат запроса или ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/firmware/v1/campaigns/{id}/pause": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Переводит кампанию из статуса Active в Paused. Выдача новых ссылок на обновление останавливается.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rollout Campaigns"
                ],
                "summary": "Приостановить кампанию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID кампании",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Кампания успешно приостановлена"
                    },
                    "400": {
                        "description": "Неверный ID кампании",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/firmware/v1/campaigns/{id}/start": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Переводит кампанию в статус Active. ESP32-устройства начинают получать обновления.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rollout Campaigns"
                ],
                "summary": "Запустить кампанию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID кампании",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Кампания успешно запущена"
                    },
                    "400": {
                        "description": "Неверный ID кампании",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/firmware/v1/campaigns/{id}/targets": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Добавляет список контроллеров ESP32 в кампанию. Доступно только для кампаний в статусе Pending.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rollout Campaigns"
                ],
                "summary": "Добавить устройства (Таргеты)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID кампании",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Список ID контроллеров",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presentation.AddTargetsRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Устройства успешно добавлены"
                    },
                    "400": {
                        "description": "Неверный формат запроса или ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_presentation.AddTargetsRequest": {
            "type": "object",
            "properties": {
                "controller_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "internal_presentation.CancelCampaignRequest": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string"
                }
            }
        },
        "internal_presentation.CreateCampaignRequest": {
            "type": "object",
            "properties": {
                "firmware_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "internal_presentation.IDResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
    }
}